Minor changes to the vdbe so that more tests pass.

+++ src/util.c
@@ -1148,7 +1148,7 @@ int sqlite3GetVarint(const unsigned char *p, u64 *v){
   u64 x = p[0] & 0x7f;
   int n = 0;
   while( (p[n++]&0x80)!=0 ){
-    x |= (p[n]&0x7f)<<(n*7);
+    x |= ((u64)(p[n]&0x7f))<<(n*7);
   }
   *v = x;
   return n;
